Fix unit tests and rename common facade helper
