+++ apiserver/common/unitcommon/package_test.go
@@ -1,7 +1,7 @@
 // Copyright 2020 Canonical Ltd.
 // Licensed under the AGPLv3, see LICENCE file for details.
 
-package caas_test
+package unitcommon_test
 
 import (
 	"testing"
