+++ apiserver/facades/agent/uniter/uniter.go
@@ -16,8 +16,8 @@ import (
 	"github.com/juju/names/v4"
 
 	"github.com/juju/juju/apiserver/common"
-	caasapi "github.com/juju/juju/apiserver/common/caas"
 	"github.com/juju/juju/apiserver/common/cloudspec"
+	"github.com/juju/juju/apiserver/common/unitcommon"
 	"github.com/juju/juju/apiserver/facade"
 	leadershipapiserver "github.com/juju/juju/apiserver/facades/agent/leadership"
 	"github.com/juju/juju/apiserver/facades/agent/meterstatus"
@@ -179,7 +179,7 @@ func NewUniterAPI(context facade.Context) (*UniterAPI, error) {
 		return nil, errors.Trace(err)
 	}
 
-	accessUnit := caasapi.CAASUnitAccessor(authorizer, caasapi.Backend(st))
+	accessUnit := unitcommon.UnitAccessor(authorizer, unitcommon.Backend(st))
 	accessApplication := applicationAccessor(authorizer, st)
 	accessMachine := machineAccessor(authorizer, st)
 	accessCloudSpec := cloudSpecAccessor(authorizer, st)
@@ -304,7 +304,7 @@ func NewUniterAPIV11(context facade.Context) (*UniterAPIV11, error) {
 	authorizer := context.Auth()
 	st := context.State()
 	resources := context.Resources()
-	accessUnit := caasapi.CAASUnitAccessor(authorizer, caasapi.Backend(st))
+	accessUnit := unitcommon.UnitAccessor(authorizer, unitcommon.Backend(st))
 	return &UniterAPIV11{
 		LXDProfileAPI: NewExternalLXDProfileAPI(st, resources, authorizer, accessUnit, logger),
 		UniterAPIV12:  *uniterAPI,
